+++ src_new/triggers/levi_ataxia/for_levi/leviticus/mnemosyne/078_Audit_Records.lua
@@ -3,6 +3,8 @@ type: trigger
 name: Mnemosyne Audit Records
 hierarchy:
 - Levi_Ataxia
+- For Levi
+- leviticus
 - Ataxia
 - Mnemosyne
 attributes: